requirements.txt
torch
torchvision
sentence-transformers
qdrant-client
numpy
tqdm
Pillow
